New option allowing processing using multiple threads
Introduces new option NUM_PROC_THREADS.
It specifies the number threads doxygen is allowed to use during
processing. When set to 0 doxygen will based this on the number of cores
available in the system. You can set it explicitly to a value larger than 0
to get more control over the balance between CPU load and processing speed.

At this moment only the input processing can be done using multiple threads.
I plan to extend this with more parallel processing in the future.

Since this is still an experimental feature the default is set to 1,
which efficively disables parallel processing. Please report any issues you
encounter that appear when changing the default.

Note that generating dot graphs in parallel is still controlled separately
by the DOT_NUM_THREADS setting.