Implement better solution for ThreadPool and std::packaged_task
Careful analysis of the compilation problems with Visual Studio revealed
that moving a std::packaged_task<R()> into a std::packaged_task<void()>
is a bit controversal as to whether or not it is supposed to work at
all or if it triggers undefined behavior.

I've now implemented the second solution as suggested here:
https://stackoverflow.com/a/31078143/784672
This avoid such conversion altogether and make the Threadpool class more generic
again.